+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `Machine.env`: Easily get the value of an environment variable
 - Allow specifying a command when spawning a subshell.
+- `ub.bootlog` and `lnx.bootlog` to allow accessing the bootlog (#5)
 
 
 ## [0.6.1] - 2018-11-16